chatbot/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
sqlalchemy>=2.0.9
psutil>=5.9.0
dingtalk-stream>=0.1.0
pymysql>=1.0.3
pydantic>=1.10.7
requests>=2.28.2
python-dotenv>=1.0.0
pyyaml>=6.0 